split MultiplicityValueProducer* in separate headers
